deepen: merge enhanced-form-parser into form-parser
Replaces 2 shallow Form parser modules with one deep Form parser module
behind a 4-method interface (parseFormContent, parseFormContentEnhanced,
extractImportantContent, extractImportantContentEnhanced).

Locality: all Form parsing — standard, URL-based, and inline-XBRL — now
lives in lib/parsers/form-parser.ts. Maintainers no longer have to bounce
between two files to follow the dispatch path.

Leverage: callers that previously had to pick between two near-identical
imports now reach for one. Behavior preserved verbatim.

Dependency category (per process/Deepening/deepening.md): in-process. No
adapter needed; the absorbed wrapper called sibling parser modules and a
.js integration module, all of which remain in-tree.

Tests deleted: 0 (the old module had no unit tests of its own).
Tests updated: 5 jest.mock paths + import paths pivoted to form-parser.
The monitor-sec-filings test's factory mock now uses jest.requireActual
to preserve the non-enhanced exports.

Stale .js shadow (lib/parsers/enhanced-form-parser.js) deleted alongside
the .ts source; it had no external importers.
