+++ __tests__/edge-cases/external-failures.test.ts
@@ -10,7 +10,7 @@ import { GET } from '../../app/api/cron/route';
 import { getPrismaClient } from '../../lib/db/prisma';
 import * as tickerMonitoring from '../../lib/sec-edgar/ticker-monitoring';
 import * as enhancedFetch from '../../lib/network/enhanced-fetch';
-import * as formParser from '../../lib/parsers/enhanced-form-parser';
+import * as formParser from '../../lib/parsers/form-parser';
 import * as summaryService from '../../services/filing/summaryGenerationService';
 import * as emailService from '../../lib/email/summary-service';
 import { logger } from '../../lib/logging';
@@ -19,7 +19,7 @@ import { logger } from '../../lib/logging';
 jest.mock('../../lib/db/prisma');
 jest.mock('../../lib/sec-edgar/ticker-monitoring');
 jest.mock('../../lib/network/enhanced-fetch');
-jest.mock('../../lib/parsers/enhanced-form-parser');
+jest.mock('../../lib/parsers/form-parser');
 jest.mock('../../services/filing/summaryGenerationService');
 jest.mock('../../lib/email/summary-service');
 jest.mock('../../lib/logging');
